Temporary solution to handle multi-digit numbers and fractions.
- Updated splitter from .split("") to .split("\\s"). "\\s" is a regex to split at whitespaces.

Note: This is only a temporary solution to a bigger issue. I feel a better parser will be required at some point. However for now, it works fairly alright.

If such a solution is to be used moving onwards to the GUI, whenever a key is pressed one will have to check the previous entered character. E.g if number clicked and number previously entered, do not add whitespace. Else if Mathematical operator entered and number previously entered, add whitespace.

(As you can see a better parser would be more ideal opposed to this means of regex splitting. It is not resilient.)