Started work on the view.
Additions
- StandardCalculator.java
The initial (potentially only) calculator type associated with this application.
- AppTheme.java
A utility class for managing the layout for the calculator. Given it is implemented through the AppTheme.set() command.
- Builder.java
A utility class for a means of quickly building multiples of GUI components with near identical similarity. Just a means of handling the process of building GUI components with the goal of reducing code considerably.

Note:
Started work on the view for the calculator. Even if we do not seek to implement functionality, such as other types of calculators, the ability to do so was still given thought to.

StandardCalculator.create() will return a StackPane for a StandardCalculator. You can imagine like that of the Windows calculator, one could implement multiple different types of calculators such as the StandardCalculator class and switch between them in the driver code by swapping StackPanes.